+++ cpsycle/dsp/src/operations.h
@@ -5,12 +5,13 @@
 #define DSP_OPERATIONS_H
 
 #include "dsptypes.h"
+#include "../../detail/stdint.h"
 
-void dsp_add(amp_t *src, amp_t *dst, int num, amp_t vol);
-void dsp_mul(amp_t *dst, int numSamples, amp_t mul);
-void dsp_movmul(amp_t *src, amp_t *dst, int num, amp_t mul);
-void dsp_clear(amp_t *dst, int num);
-void dsp_interleave(amp_t* dst, amp_t* left, amp_t* right, int num);
-void dsp_erase_all_nans_infinities_and_denormals(amp_t* dst, unsigned int num);
+void dsp_add(amp_t *src, amp_t *dst, uintptr_t num, amp_t vol);
+void dsp_mul(amp_t *dst, uintptr_t numSamples, amp_t mul);
+void dsp_movmul(amp_t *src, amp_t *dst, uintptr_t num, amp_t mul);
+void dsp_clear(amp_t *dst, uintptr_t num);
+void dsp_interleave(amp_t* dst, amp_t* left, amp_t* right, uintptr_t num);
+void dsp_erase_all_nans_infinities_and_denormals(amp_t* dst, uintptr_t num);
 
 #endif
